homepage, 404 page, added additional files
