+++ client/src/components/Navbar.js
@@ -32,14 +32,14 @@ const AppNavbar = () => {
                             <Nav.Link as={Link} to='/about'>
                                 About
                             </Nav.Link>
-                            <Nav.Link as={Link} to='/Coin'>
+                            <Nav.Link as={Link} to='/currencySearch'>
                                 Coin Stats
                             </Nav.Link>
                             {/* if user is logged in show saved books and logout */}
                             {Auth.loggedIn() ? (
                                 <>
-                                    <Nav.Link as={Link} to='/saved'>
-                                        See Your Books
+                                    <Nav.Link as={Link} to='/profile'>
+                                        Profile
                                     </Nav.Link>
                                     <Nav.Link onClick={Auth.logout}>Logout</Nav.Link>
                                 </>
