+++ client/src/components/pages/About.js
@@ -1,0 +1,24 @@
+
+
+const About = () => {
+    return (
+        <div>
+            <h1>About Page</h1>
+            <p>
+                Welcome to Crypto Coin, the go-to website for all things related to cryptocurrency news and updates. Our website offers a comprehensive platform where users can stay up-to-date with the latest trends, insights, and news related to cryptocurrency.
+
+                At Crypto Coin, we believe that knowledge is power, and that's why we offer three different subscription tiers to cater to users' needs. Whether you're a beginner, intermediate, or advanced investor, our subscription plans are designed to provide customized news and updates to meet your specific requirements.
+
+                Our team of experts curates and compiles relevant news and insights from the world of cryptocurrency to provide our users with accurate and insightful information. Our subscription plans come with a range of features, including real-time news alerts, customized research reports, expert analysis, and more.
+
+                Our beginner subscription plan is perfect for those who are new to the world of cryptocurrency and want to learn more about the basics. The intermediate plan is designed for those who have some knowledge of cryptocurrency and want to deepen their understanding. And for the advanced investor, our subscription plan offers in-depth analysis and insights to help them make informed investment decisions.
+
+                We believe that our subscription plans offer immense value to our users by providing them with the latest news and insights related to cryptocurrency. With Crypto Coin, you can stay ahead of the game and make well-informed decisions.
+
+                Thank you for choosing Crypto Coin. We look forward to providing you with the latest news and insights related to cryptocurrency.
+            </p>
+        </div>
+    );
+}
+
+export default About;
