+++ package.json
@@ -7,7 +7,7 @@
     "start": "npm run build && cd server && node server.js",
     "server": "cd server nodemon server.js --ignore client",
     "client": "cd client && npm start",
-    "develop": "concurrently \"cd server && npm run server\" \"cd client && npm run dev\"",
+    "develop": "concurrently \"cd server && npm run server\" \"cd client && npm start\"",
     "install": "cd server && npm i && cd ../client && npm i",
     "seed": "cd server && npm run seed",
     "build": "cd client && npm run build",
